Feature/change layer power and speed to double (#306)
* change layer speed and power to double

* Add error message when the bb file version overrange

+++ src/settings/preset-settings.cpp
@@ -3,8 +3,8 @@
 PresetSettings::Param PresetSettings::Param::fromJson(const QJsonObject &obj) {
   Param p;
   p.name = obj["name"].toString();
-  p.power = obj["power"].toInt();
-  p.speed = obj["speed"].toInt();
+  p.power = obj["power"].toDouble();
+  p.speed = obj["speed"].toDouble();
   p.repeat = obj["repeat"].toInt(1);
   p.step_height = obj["step_height"].toDouble();
   p.target_height = obj["target_height"].toDouble();
